Enum to track audiosource state. Audio accessors. Reorganised folder structure.
