+++ Assets/Prefabs/UI/UIManager.prefab
@@ -63,7 +63,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7513093811224638330}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5864ce12624ca1b4a9714c856dd231a8, type: 3}
+  m_Script: {fileID: 11500000, guid: 5e81190e38f7c8e4ab867b26d752c122, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   bedCanvas: {fileID: 8131493421899459369}
