+++ Assets/Scripts/Extensions/ObjectExtensions.cs
@@ -1,0 +1,22 @@
+﻿using UnityEngine;
+
+public static class ObjectExtensions
+{
+    public static void LogIfNull(this Object self, string objectDescription = null)
+    {
+        if (self != null)
+        {
+            return;
+        }
+
+        string textToLog = string.Empty;
+
+        if (!string.IsNullOrEmpty(objectDescription))
+        {
+            textToLog += objectDescription + " ";
+        }
+
+        textToLog += $"Object of type {self.GetType()} is null";
+        Debug.LogError(textToLog);
+    }
+}
