+++ Assets/Scripts/Singletons/AudioSingletons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 570888908b179b14fa8d7f3e2495cbda
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
