+++ Assets/Scripts/Singletons/AudioSingletons/MusicManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5864ce12624ca1b4a9714c856dd231a8
+guid: aef478e06d6c26642bdd5f7bc424aad0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
