+++ Assets/Scripts/Singletons/UISingletons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1852c6864ad7be44aab65dcae77c884
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
