+++ Assets/Scripts/Singletons/UISingletons/FontManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cc51cdb13ca51ce408a70f69bf35e607
+guid: 38528ad2c594d2f478e5e9da3ddc9863
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
