+++ Assets/Scripts/Singletons/UISingletons/UIManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba033483a919926478135bbafbb062f0
+guid: 5e81190e38f7c8e4ab867b26d752c122
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
